==> notebooks/04_model_training.ipynb <==
PURPOSE OF THIS NOTEBOOK:
This notebook is used to train models using X_train_final and y_train

from sklearn.linear_model import LogisticRegression
from scipy.sparse import load_npz
import numpy as np
#Logistic Regression is a model that estimates the probability that something belongs in a class
# In this case, what is the probability that this is fraudlent?

model=LogisticRegression(max_iter=1000,class_weight="balanced")
#Without using balanced, model can predict all jobs are 0 that is real and still get ~ 95% accurancy which is useless

X_train_final = load_npz("../data/processed/X_train_final.npz")
X_test_final = load_npz("../data/processed/X_test_final.npz")

y_train = np.load("../data/processed/y_train.npy")
y_test = np.load("../data/processed/y_test.npy")

model.fit(X_train_final,y_train)

y_pred=model.predict(X_test_final)
