==> notebooks/05_model_evaluation.ipynb <==
PURPOSE OF THIS NOTEBOOK
This notebook is to check if the model is actually detecting fraud or not.

import joblib
from scipy.sparse import load_npz
import numpy as np
# Save trained model
model=joblib.load("../models/logistic_regression.pkl")

X_train_final = load_npz("../data/processed/X_train_final.npz")
X_test_final = load_npz("../data/processed/X_test_final.npz")

y_train = np.load("../data/processed/y_train.npy")
y_test = np.load("../data/processed/y_test.npy")

from sklearn.metrics import confusion_matrix
y_pred=model.predict(X_test_final)
cm=confusion_matrix(y_test,y_pred)
cm

from sklearn.metrics import classification_report
print(classification_report(y_test,y_pred))

from sklearn.metrics import f1_score,precision_score, recall_score
precision = precision_score(y_test,y_pred)
recall = recall_score(y_test,y_pred)
f1 = f1_score(y_test,y_pred)
precision, f1,recall

from sklearn.metrics import classification_report

# Training performance
y_train_pred = model.predict(X_train_final)
print("TRAIN METRICS")
print(classification_report(y_train, y_train_pred))

# Test performance
print("TEST METRICS")
print(classification_report(y_test, y_pred))
